requirements.txt
asgiref==3.5.2
boto3==1.24.89
botocore==1.27.89
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.1.1
PyJWT==2.5.0
python-dateutil==2.8.2
pytz==2022.4
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
